requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncache==0.1.1
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
charset-normalizer==2.1.1
discord.py==2.4.0
frozenlist==1.3.1
fuzzywuzzy==0.18.0
hijri-converter==2.2.4
html2text==2020.1.16
html5lib==1.1
idna==3.3
multidict==6.0.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.2.1
ratelimiter==1.2.0.post0
six==1.16.0
soupsieve==2.3.2.post1
topggpy==1.4.0
webencodings==0.5.1
yarl==1.8.1
